Rip main::EvalError out into err::EvalError
